// template/react-starter-kit/src/.generated/api/model/termsControllerGetTermHistoryCursor200.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { GetTermHistoryCursorResponseDto } from './getTermHistoryCursorResponseDto';
import type { TermsControllerGetTermHistoryCursor200Meta } from './termsControllerGetTermHistoryCursor200Meta';

export type TermsControllerGetTermHistoryCursor200 = {
  success: boolean;
  statusCode: number;
  path: string;
  requestId: string;
  timestamp: string;
  data: GetTermHistoryCursorResponseDto;
  message?: string;
  meta?: TermsControllerGetTermHistoryCursor200Meta;
};